Add Node.js Assignment 3

+++ src/service/todos.ts
@@ -1,17 +1,18 @@
+import { todos } from "../data/todos";
+import { NotFoundError } from "../error/Errors";
 import { ITodo } from "../interface/todo";
 import * as TodoModel from "../model/todos"
-export function fetchTodos(){
-    const todos = TodoModel.fetchTodos();
+import { ROLES } from "../utils/enum";
+
+export function fetchTodos(userId: string, role: ROLES){
+    const todos = TodoModel.fetchTodos(userId, role);
+    if(!todos.length) throw new NotFoundError(`No todos found`);
     return todos;
 }
 
-export function fetchTodoById(id: string){
-    const result = TodoModel.fetchTodoById(id);
-    if(!result){
-        return {
-            error: `Todo with id ${id} does not exist.`
-        }
-    }
+export function fetchTodoById(id: string, role: ROLES, userId: string){
+    const result = TodoModel.fetchTodoById(id, role, userId);
+    if(!result) throw new NotFoundError(`Todo with id ${id} does not exist`)
     return result;
 }
 
@@ -22,13 +23,25 @@ export function addTodo(todo: Pick<ITodo, "userId" | "title" | "description" >){
 }
 
 export function deleteTodoById(id: string){
+    const todo = todos.find(todo => todo.id === id);
+    if(!todo) throw new NotFoundError(`Todo with id ${id} does not exist`);
     return TodoModel.deleteTodoById(id);
 }
 
 export function updateTodo(id: string, todo: ITodo){
+    const existingTodo = todos.find(todo => todo.id === id);
+    if(!existingTodo) throw new NotFoundError(`Todo with id ${id} does not exist`);
     return TodoModel.updateTodo(id, todo);
 }
 
 export function finishTask(id: string){
+    const existingTodo = todos.find(todo => todo.id === id);
+    if(!existingTodo) throw new NotFoundError(`Todo with id ${id} does not exist`);
     return TodoModel.finishTask(id);
+}
+
+export function fetchFinishedTask(userId: string){
+    const finishedTask =  TodoModel.fetchFinishedTask(userId);
+    if(!finishedTask.length) throw new NotFoundError(`No finished todos.`);
+    return finishedTask;
 }
